Clean up minor warnings and code style
* Remove warnings for code in vendor folders
* Add pedantic and extra warnings in GCC only to projectM sources
* Rename ResetOpenGL to SetWindowSize

+++ src/libprojectM/TimeKeeper.cpp
@@ -95,12 +95,12 @@ int TimeKeeper::PresetFrameA()
     return m_presetFrameA;
 }
 
-int TimeKeeper::PresetTimeB()
+double TimeKeeper::PresetTimeB()
 {
     return m_presetTimeB;
 }
 
-int TimeKeeper::PresetTimeA()
+double TimeKeeper::PresetTimeA()
 {
     return m_presetTimeA;
 }
